# docs/swagger.yaml
basePath: /
...
host: localhost:8000
info:
  contact: {}
  description: This is the API documentation for the Appointment App.
  title: Appointment App API
  version: "1.0"
paths:
...
  /providers/{id}/review-stats:
    get:
      consumes:
      - application/json
      description: Retrieves review statistics for a specific provider
      parameters:
      - description: Provider ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              AvgRating:
                type: number
              ProviderID:
                type: integer
              Rating1Count:
                type: integer
              Rating2Count:
                type: integer
              Rating3Count:
                type: integer
              Rating4Count:
                type: integer
              Rating5Count:
                type: integer
              TotalReviews:
                type: integer
            type: object
        "500":
          description: Internal Server Error
          schema:
            properties:
              error:
                type: string
            type: object
      summary: Get provider review statistics
      tags:
      - Reviews
  /providers/{id}/reviews:
    get:
      consumes:
      - application/json
      description: Retrieves a paginated list of reviews for a specific provider
      parameters:
      - description: Provider ID
        in: path
        name: id
        required: true
        type: string
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Number of items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              limit:
                type: integer
              page:
                type: integer
              pages:
                type: integer
              reviews:
                items:
                  properties:
                    AppointmentID:
                      type: integer
                    Comment:
                      type: string
                    CreatedAt:
                      type: string
                    Customer:
                      properties:
                        CreatedAt:
                          type: string
                        ID:
                          type: integer
                        Name:
                          type: string
                      type: object
                    ID:
                      type: integer
                    IsAnonymous:
                      type: boolean
                    IsVerified:
                      type: boolean
                    ProviderID:
                      type: integer
                    Rating:
                      type: number
                    Service:
                      properties:
                        ID:
                          type: integer
                        Name:
                          type: string
                      type: object
                    UpdatedAt:
                      type: string
                  type: object
                type: array
              total:
                type: integer
            type: object
        "500":
          description: Internal Server Error
          schema:
            properties:
              error:
                type: string
            type: object
      summary: Get provider reviews
      tags:
      - Reviews
...
  /reviews:
    post:
      consumes:
      - application/json
      description: Creates a new review for a provider
      parameters:
      - description: Review details
        in: body
        name: review
        required: true
        schema:
          properties:
            AppointmentID:
              type: integer
            Comment:
              type: string
            IsAnonymous:
              type: boolean
            ProviderID:
              type: integer
            Rating:
              type: number
            ServiceID:
              type: integer
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            properties:
              AppointmentID:
                type: integer
              Comment:
                type: string
              CreatedAt:
                type: string
              CustomerID:
                type: integer
              ID:
                type: integer
              IsAnonymous:
                type: boolean
              IsVerified:
                type: boolean
              ProviderID:
                type: integer
              Rating:
                type: number
              ServiceID:
                type: integer
              UpdatedAt:
                type: string
            type: object
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            properties:
              error:
                type: string
            type: object
        "404":
          description: Not Found
          schema:
            properties:
              error:
                type: string
            type: object
        "409":
          description: Conflict
          schema:
            properties:
              error:
                type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create a review
      tags:
      - Reviews
  /reviews/{id}:
    delete:
      consumes:
      - application/json
      description: Deletes a review by its ID
      parameters:
      - description: Review ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            properties:
              error:
                type: string
            type: object
        "403":
          description: Forbidden
          schema:
            properties:
              error:
                type: string
            type: object
        "404":
          description: Not Found
          schema:
            properties:
              error:
                type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete a review
      tags:
      - Reviews
    put:
      consumes:
      - application/json
      description: Updates a review by its ID
      parameters:
      - description: Review ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated review details
        in: body
        name: review
        required: true
        schema:
          properties:
            Comment:
              type: string
            IsAnonymous:
              type: boolean
            Rating:
              type: number
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              AppointmentID:
                type: integer
              Comment:
                type: string
              CreatedAt:
                type: string
              CustomerID:
                type: integer
              ID:
                type: integer
              IsAnonymous:
                type: boolean
              IsVerified:
                type: boolean
              ProviderID:
                type: integer
              Rating:
                type: number
              ServiceID:
                type: integer
              UpdatedAt:
                type: string
            type: object
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            properties:
              error:
                type: string
            type: object
        "403":
          description: Forbidden
          schema:
            properties:
              error:
                type: string
            type: object
        "404":
          description: Not Found
          schema:
            properties:
              error:
                type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update a review
      tags:
      - Reviews
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"